Former-commit-id: 4cfc703e96dea0318ac8e5e815958997b1c3fb97

+++ modules/sofa/component/mapping/RigidMapping.inl
@@ -1022,8 +1022,8 @@ const sofa::defaulttype::BaseMatrix* RigidMapping<TIn, TOut>::getJ()
     {
         updateJ = false;
         if (matrixJ.get() == 0 ||
-                matrixJ->rowBSize() != out.size() ||
-                matrixJ->colBSize() != in.size())
+                (unsigned int)matrixJ->rowBSize() != out.size() ||
+                (unsigned int)matrixJ->colBSize() != in.size())
         {
             matrixJ.reset(new MatrixType(out.size() * NOut, in.size() * NIn));
         }
